Create pingsweeper.py
A ping sweep, also known as an ICMP sweep is a basic network scanning technique which is used to determine the ranges of IP addresses that maps to live hosts(systems). Whereas, a single ping will tell you whether one specified host computer exists on the network, a ping sweep consists of ICMP ECHO requests, sent to each IP address to see if it is live. If the given address is live, it will return an ICMP ECHO reply. Ping sweeps are among the older and slower methods used to scan the network.

The above code is a python script used for finding live hosts by using the ping sweep.